% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_disc.R
\name{poisson_disc}
\alias{poisson_disc}
\title{Poisson disc sampling}
\usage{
poisson_disc(width, height, min_dist, k = 30)
}
\arguments{
\item{width}{The canvas width (numeric)}

\item{height}{The canvas height (numeric)}

\item{min_dist}{The minimum distance between points (numeric)}

\item{k}{Samples to choose before rejection (numeric), this should usually stay at 30}
}
\value{
A data frame with the x and y coordinates of the points.
}
\description{
Generates a set of (x, y) points that are never closer than some minimum distance in a canvas of given width and height.
}
\examples{
#generate a set of points within a 400x400 canvas with a minimum distance of 20
pts <- poisson_disc(400, 400, 20)


}
